back_end/requirements.txt
Flask==2.0.2
PyJWT==2.3.0